Rename summary_statistics -> summary_stats Name changes made for consistency, as suggested by @hayfield
