+++ static/templates/summary_stats.html
@@ -14,7 +14,7 @@
 
     <div class="panel panel-default" id="h_table">
         <div class="panel-heading">
-            <span class="pull-right"><a href="{{url('summary_statistics.csv')}}">(This table as CSV)</a></span>
+            <span class="pull-right"><a href="{{url('summary_stats.csv')}}">(This table as CSV)</a></span>
             <h3 class="panel-title">Summary Statistics</h3>
         </div>
 
@@ -31,16 +31,16 @@
             <thead>
                 <tr>
                     <th style="vertical-align:middle; border: 1px solid gray;">Publisher Name</th>
-                    {% for column_slug, column_header in summary_statistics.columns %}
+                    {% for column_slug, column_header in summary_stats.columns %}
                     <th style="vertical-align:middle; border: 1px solid gray;">{{column_header}}</th>
                     {% endfor %}
                 </tr>
             </thead>
             <tbody>
-                {% for row in summary_statistics.table() %}
+                {% for row in summary_stats.table() %}
                 <tr {% if loop.last %} style="border-bottom: 1px solid gray;"{% endif %}>
                     <td style="border-right: 1px solid gray; border-left: 1px solid gray;"><a href="publisher/{{row.publisher}}.html">{{row.publisher_title}}</a></td>
-                    {% for column_slug, column_header in summary_statistics.columns %}
+                    {% for column_slug, column_header in summary_stats.columns %}
                     <td style="border-right: 1px solid gray; border-left: 1px solid gray;">{{row[column_slug]}}</th>
                     {% endfor %}
                     </td>
@@ -154,10 +154,10 @@
         <div class="panel-body">
             <h5>Secondary reporters</h5>
             <p>Publishers who publish all of their activities as a secondary reporter do not appear in this table.
-            {% if summary_statistics.secondary_publishers|length > 0 %}
-            Therefore, a total of {{ summary_statistics.secondary_publishers|length }} publishers have been excluded in this regard:
+            {% if summary_stats.secondary_publishers|length > 0 %}
+            Therefore, a total of {{ summary_stats.secondary_publishers|length }} publishers have been excluded in this regard:
             <ul>
-            {% for publisher in summary_statistics.secondary_publishers %}
+            {% for publisher in summary_stats.secondary_publishers %}
                 <li>{{ publisher_name[publisher] }}</li>
             {% endfor %}
             </ul>
